refactor: :label: added types file for shared types
