+++ src/types.ts
@@ -1,0 +1,10 @@
+export type User = {
+  id: string;
+  username: string;
+};
+
+export type Task = {
+  id: string;
+  text: string;
+  imageId: string;
+};
